chore: fix clippy warnings and apply cargo fmt
Resolve all clippy::all + clippy::pedantic warnings across the codebase:

- Replace deprecated criterion::black_box with std::hint::black_box
- Fix manual_checked_ops: integer division replaced with checked_div
- Fix needless_range_loop: index loops converted to iterator patterns
- Fix ptr_arg: &mut Vec<T> params changed to &mut [T] at call boundaries
- Fix field_reassign_with_default: use struct update syntax
- Fix erasing_op and approx_constant (deny-level lints)
- Fix assertions_on_constants in tests
- Fix doc lints and remove unused imports throughout
- Add targeted #[allow(...)] with reason comments for intentional lints
  (too_many_arguments and needless_range_loop in SIMD/GPU kernels,
  type_complexity and wrong_self_convention at PyO3 boundaries,
  should_implement_trait on FeatureExpr operators, dead_code on
  test helpers)

Apply cargo fmt --all formatting pass.

+++ src/serialize/trb.rs
@@ -1163,7 +1163,7 @@ mod tests {
         file.read_exact(&mut total_size_bytes).unwrap();
         let mut header_size_bytes = [0u8; 8];
         file.read_exact(&mut header_size_bytes).unwrap();
-        let header_size = u64::from_le_bytes(header_size_bytes) as u64;
+        let header_size = u64::from_le_bytes(header_size_bytes);
 
         // Blob starts at: base_end + 8 (total_size) + 8 (header_size) + header_size
         let blob_start = base_end + 8 + 8 + header_size;
